fix: incomplete reads of long IMAP messages
Previously, the `ImapStream::poll_next` method did not persist its state
correctly across `Poll::Pending` results of the underlying byte stream.
This resulted in an empty buffer and empty position being used for each
following call of `poll_next`, which therefore caused the `decode`
method being invoked not on the entire accumulated buffer, but on
individual chunks of input.

This commit fixes this problematic behavior by properly persisting
the buffer and the current position across `poll_next` invocations.
In addition, this commit replaces the (usize, usize) structure
representing position with a proper struct, which makes more
semantically and syntactically clear which kind of index is being used.