Ensure realloced blocks are aligned
The blocks stay around in the byte-pool forever, when asking
non-aligned blocks each size will result in a new allocation even if a
previous block would fit.  This improves this situation by always
asking for blocks multiples of INITIAL_CAPACITY which should increase
the likelyhood that a block is re-used a lot.

Thoug larger unused blocks will still not be re-used when they could,
which is arguably fine.