Do not generate artificial "broken pipe" errors
`closed` flag is set when there is no more data to read from the stream.
It does not mean that no more requests can be written into the stream.
If the stream is closed for writing, `write_all` call
will return this error, there is no need to emulate it.