Fix setting of decode_needs variable
Previously, `decode_needs` was set to the value
returned by `nom` in the `nom::Err::Incomplete` error.
This is incorrect, because `nom` returns the number
of additional bytes needed rather than the total number of bytes.
Current filled buffer size should be added to this
number to obtain `decode_needs` value.

In addition, unwrap `decode_needs` from `Option`.
`None` was always unwrapped with `unwrap_or(0)`,
so we can as well store 0 directly and save the memory.