fix: reset IDLE timeout when keepalive is received
This allows to use shorter timeouts
to detect lost connection earlier.
E.g. if timeout is set to 5 minutes
and the server sends keepalives every 2 minutes,
IDLE will never be interrupted.
If connection is lost, it will be noticed not later than 5 minutes later.

`wait` method which does not accept the timeout
now uses default timeout of 24 hours,
but should probably be deprecated.